from int to long

+++ app/src/main/java/com/example/palfinder/Person.kt
@@ -1,7 +1,7 @@
 package com.example.palfinder
 
 class Person(var name: String? = null,
-             var socialSecurityNumber : Int? = null,
+             var socialSecurityNumber : Long? = null,
              var city :String? = null,
              var phoneNumber : String? = null,
              var email : String? = null,
